+++ package.json
@@ -26,7 +26,8 @@
     "express": "^4.16.4",
     "joi": "^14.3.1",
     "joi-date-extensions": "^1.2.0",
-    "morgan": "^1.9.1"
+    "morgan": "^1.9.1",
+    "pg": "^7.8.0"
   },
   "devDependencies": {
     "chai": "^4.2.0",
