+++ server/routes/meetup.js
@@ -1,13 +1,15 @@
 const express = require('express');
 
 const app = express.Router();
+const meetupcontrollers = require('../controllers/meetup');
 
-const userController = require('../controllers/meetup');
+// app.get("/", meetupcontrollers.getMeetup);
+// app.post("/", meetupcontrollers.registerMeetup);
 
-app.post('/api/v1/meetups', userController.registerM);
-app.get('/api/v1/meetups', userController.allmeetup);
-app.get('/api/v1/meetups/:id', userController.meetupId);
-app.patch('/api/v1/meetups/:id', userController.updateMeetup);
-app.delete('/api/v1/meetups/:id', userController.deleteMeetup);
+// app.post('/api/v1/meetups', userController.registerM);
+app.get('/', meetupcontrollers.getMeetup);
+// app.get('/api/v1/meetups/:id', userController.meetupId);
+// app.patch('/api/v1/meetups/:id', userController.updateMeetup);
+// app.delete('/api/v1/meetups/:id', userController.deleteMeetup);
 
 module.exports = app;
